up 9 mai

+++ master/data/src/main/java/com/wacajou/data/jpa/repository/ModuleRepository.java
@@ -12,7 +12,7 @@ import com.wacajou.data.jpa.domain.ParcoursModule;
 import com.wacajou.data.jpa.domain.User;
 import com.wacajou.data.jpa.domain.UserModule;
 
-public interface ModuleRepository  extends JpaRepository<Module, Integer>{
+public interface ModuleRepository  extends JpaRepository<Module, Long>{
 	
 	public static final String FIND_BY_USER = "SELECT m " + 
 			"FROM Module m " +
